More WrenHost into shared base class
