+++ Source/Client/Core/Source/Module.cpp
@@ -21,6 +21,7 @@ module;
 #include <sstream>
 #include <stack>
 #include <string>
+#include <fstream>
 #include <unordered_map>
 #include <unordered_set>
 #include <variant>
@@ -33,16 +34,22 @@ module;
 #undef min
 #undef CreateProcess
 #undef GetCurrentTime
+#undef GetClassName
 
 #endif
 
 #ifdef SOUP_BUILD
+
+// TODO module
+#include "Wren.hpp"
+
 export module Soup.Core;
 
 import reflex;
 import CryptoPP;
 import Monitor.Host;
 import Opal;
+
 
 using namespace Opal;
 
@@ -84,6 +91,9 @@ using namespace Opal;
 // import Monitor.Host
 #include "Linux/LinuxMonitorProcessManager.h"
 
+// import Wren
+#include "Wren.hpp"
+
 #endif
 
 #define CLIENT_CORE_IMPLEMENTATION
@@ -92,3 +102,5 @@ using namespace Opal;
 #include "Build/BuildEngine.h"
 #include "LocalUserConfig/LocalUserConfigExtensions.h"
 #include "Package/PackageManager.h"
+#include "Wren/WrenHost.h"
+#include "Wren/WrenValueTable.h"
