+++ Source/Generate/ExtensionManager.h
@@ -4,7 +4,7 @@
 
 #pragma once
 #include "ExtensionTaskDetails.h"
-#include "WrenHost.h"
+#include "GenerateHost.h"
 
 namespace Soup::Core::Generate
 {
@@ -111,7 +111,7 @@ namespace Soup::Core::Generate
 					throw std::runtime_error("TryFindNextTask returned empty result");
 
 				// Create a Wren Host to evaluate the extension task
-				auto host = std::make_unique<WrenHost>(currentTask->ScriptFile);
+				auto host = std::make_unique<GenerateHost>(currentTask->ScriptFile, std::nullopt);
 				host->InterpretMain();
 
 				// Set the current state AFTER we initialize to prevent pre-loading
