+++ Source/Generate/GenerateHost.h
@@ -1,70 +1,30 @@
 #pragma once
 
-#include "wren.hpp"
-#include "WrenHelpers.h"
-#include "WrenValueTable.h"
 #include "ExtensionTaskDetails.h"
 #include "GenerateState.h"
 
 namespace Soup::Core::Generate
 {
-	class WrenHost
+	class GenerateHost : public WrenHost
 	{
 	private:
 		static inline const char* SoupModuleName = "soup";
 		static inline const char* SoupClassName = "Soup";
 		static inline const char* SoupTaskClassName = "SoupTask";
-		Path _scriptFile;
-		WrenVM* _vm;
+
+	private:
 		GenerateState* _state;
 
 	public:
-		WrenHost(Path scriptFile) :
-			_scriptFile(std::move(scriptFile)),
-			_vm(nullptr),
+		GenerateHost(Path scriptFile, std::optional<Path> bundlesFile) :
+			WrenHost(std::move(scriptFile), std::move(bundlesFile)),
 			_state(nullptr)
 		{
-			// Configure the Wren Virtual Machine
-			WrenConfiguration config;
-			wrenInitConfiguration(&config);
-			config.resolveModuleFn = &WrenResolveModule;
-			config.loadModuleFn = &WrenLoadModule;
-			config.bindForeignMethodFn = &WrenBindForeignMethod;
-			config.writeFn = &WrenWriteCallback;
-			config.errorFn = &WrenErrorCallback;
-			config.userData = this;
-
-			// Create the VM
-			_vm = wrenNewVM(&config);
-		}
-
-		~WrenHost()
-		{
-			wrenFreeVM(_vm);
-			_vm = nullptr;
 		}
 
 		void SetState(GenerateState& state)
 		{
 			_state = &state;
-		}
-
-		void InterpretMain()
-		{
-			Log::Diag("InterpretMain");
-
-			// Load the script
-			std::ifstream scriptFile(_scriptFile.ToString());
-			if (!scriptFile.is_open())
-				throw std::runtime_error("Script does not exist");
-
-			auto script = std::string(
-				std::istreambuf_iterator<char>(scriptFile),
-				std::istreambuf_iterator<char>());
-
-			// Interpret the script
-			WrenHelpers::ThrowIfFailed(wrenInterpret(_vm, _scriptFile.ToString().c_str(), script.c_str()));
-			Log::Diag("Success");
 		}
 
 		std::vector<ExtensionTaskDetails> DiscoverExtensions()
@@ -82,7 +42,7 @@ namespace Soup::Core::Generate
 				auto type = wrenGetSlotType(_vm, 0);
 				if (type == WREN_TYPE_UNKNOWN)
 				{
-					auto classHandle = WrenHelpers::SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
+					auto classHandle = SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
 					if (WrenHelpers::HasParentType(_vm, classHandle, SoupTaskClassName))
 					{
 						Log::Diag("Found Build Task");
@@ -115,10 +75,10 @@ namespace Soup::Core::Generate
 				throw std::runtime_error("Extension class name was not a class");
 			}
 
-			auto classHandle = WrenHelpers::SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
+			auto classHandle = SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
 
 			// Call Evaluate
-			auto evaluateMethodHandle = WrenHelpers::SmartHandle(_vm, wrenMakeCallHandle(_vm, "evaluate()"));
+			auto evaluateMethodHandle = SmartHandle(_vm, wrenMakeCallHandle(_vm, "evaluate()"));
 
 			wrenSetSlotHandle(_vm, 0, classHandle);
 			WrenHelpers::ThrowIfFailed(wrenCall(_vm, evaluateMethodHandle));
@@ -135,10 +95,10 @@ namespace Soup::Core::Generate
 				throw std::runtime_error("Missing Class Soup");
 			}
 
-			auto soupClassHandle = WrenHelpers::SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
+			auto soupClassHandle = SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
 
 			// Call ActiveState
-			auto activeStateGetterHandle = WrenHelpers::SmartHandle(_vm, wrenMakeCallHandle(_vm, "activeState"));
+			auto activeStateGetterHandle = SmartHandle(_vm, wrenMakeCallHandle(_vm, "activeState"));
 			wrenSetSlotHandle(_vm, 0, soupClassHandle);
 			WrenHelpers::ThrowIfFailed(wrenCall(_vm, activeStateGetterHandle));
 
@@ -167,10 +127,10 @@ namespace Soup::Core::Generate
 				throw std::runtime_error("Missing Class Soup");
 			}
 
-			auto soupClassHandle = WrenHelpers::SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
+			auto soupClassHandle = SmartHandle(_vm, wrenGetSlotHandle(_vm, 0));
 
 			// Call SharedState
-			auto sharedStateGetterHandle = WrenHelpers::SmartHandle(_vm, wrenMakeCallHandle(_vm, "sharedState"));
+			auto sharedStateGetterHandle = SmartHandle(_vm, wrenMakeCallHandle(_vm, "sharedState"));
 			wrenSetSlotHandle(_vm, 0, soupClassHandle);
 			WrenHelpers::ThrowIfFailed(wrenCall(_vm, sharedStateGetterHandle));
 
@@ -189,10 +149,68 @@ namespace Soup::Core::Generate
 		}
 
 	private:
+		virtual bool IsBuiltInModule(std::string_view moduleName) override final
+		{
+			if (moduleName == std::string_view(SoupModuleName))
+			{
+				return true;
+			}
+			else
+			{
+				return false;
+			}
+		}
+
+		virtual bool TryGetBuiltInModule(std::string_view moduleName, const char*& source) override final
+		{
+			// Inject Soup module
+			if (moduleName == SoupModuleName)
+			{
+				source = GetSoupModuleSource();
+				return true;
+			}
+			else
+			{
+				source = nullptr;
+				return false;
+			}
+		}
+
+		virtual WrenForeignMethodFn TryBindForeignMethod(
+			std::string_view moduleName,
+			std::string_view className,
+			bool isStatic,
+			std::string_view signature) override final
+		{
+			// There is only one foreign method in the soup module.
+			if (moduleName == SoupModuleName)
+			{
+				if (className == SoupClassName && isStatic)
+				{
+					if (signature == "loadGlobalState_()")
+						return SoupLoadGlobalState;
+					else if (signature == "loadActiveState_()")
+						return SoupLoadActiveState;
+					else if (signature == "loadSharedState_()")
+						return SoupLoadSharedState;
+					else if (signature == "createOperation_(_,_,_,_,_,_)")
+						return SoupCreateOperation;
+					else if (signature == "info_(_)")
+						return SoupLogInfo;
+					else if (signature == "warning_(_)")
+						return SoupLogWarning;
+					else if (signature == "error_(_)")
+						return SoupLogError;
+				}
+			}
+
+			return nullptr;
+		}
+
 		std::vector<std::string> CallRunBeforeGetter(WrenHandle* classHandle)
 		{
 			// Call RunBefore
-			auto runBeforeGetterHandle = WrenHelpers::SmartHandle(_vm, wrenMakeCallHandle(_vm, "runBefore"));
+			auto runBeforeGetterHandle = SmartHandle(_vm, wrenMakeCallHandle(_vm, "runBefore"));
 
 			wrenSetSlotHandle(_vm, 0, classHandle);
 			WrenHelpers::ThrowIfFailed(wrenCall(_vm, runBeforeGetterHandle));
@@ -203,43 +221,12 @@ namespace Soup::Core::Generate
 		std::vector<std::string> CallRunAfterGetter(WrenHandle* classHandle)
 		{
 			// Call RunAfter
-			auto runBeforeGetterHandle = WrenHelpers::SmartHandle(_vm, wrenMakeCallHandle(_vm, "runAfter"));
+			auto runBeforeGetterHandle = SmartHandle(_vm, wrenMakeCallHandle(_vm, "runAfter"));
 
 			wrenSetSlotHandle(_vm, 0, classHandle);
 			WrenHelpers::ThrowIfFailed(wrenCall(_vm, runBeforeGetterHandle));
 
 			return WrenHelpers::GetSlotStringList(_vm, 0, 1);
-		}
-
-		void WriteCallback(std::string_view text)
-		{
-			Log::Diag(text);
-		}
-
-		void ErrorCallback(
-			WrenErrorType errorType,
-			std::optional<std::string_view> moduleName,
-			int line,
-			std::string_view message)
-		{
-			switch (errorType)
-			{
-				case WREN_ERROR_COMPILE:
-				{
-					printf("[%s(%d)] [Error] %s\n", moduleName.value().data(), line, message.data());
-					break;
-				}
-				case WREN_ERROR_STACK_TRACE:
-				{
-					printf("[%s(%d)] in %s\n", moduleName.value().data(), line, message.data());
-					break;
-				}
-				case WREN_ERROR_RUNTIME:
-				{
-					printf("[Runtime Error] %s\n", message.data());
-					break;
-				}
-			}
 		}
 
 		void SoupLoadGlobalState()
@@ -373,197 +360,51 @@ namespace Soup::Core::Generate
 			wrenSetSlotNull(_vm, 0);
 		}
 
-		static WrenForeignMethodFn WrenBindForeignMethod(
-			WrenVM* vm,
-			const char* module,
-			const char* className,
-			bool isStatic,
-			const char* signature)
-		{
-			(vm);
-			(module);
-			auto classNameValue = std::string_view(className);
-			auto signatureValue = std::string_view(signature);
-
-			WrenForeignMethodFn result = nullptr;
-
-			// Attempt to bind to the Soup Module
-			result = TryBindSoupModuleForeignMethod(classNameValue, isStatic, signatureValue);
-			if (result != nullptr)
-				return result;
-
-			return nullptr;
-		}
-
-		static const char* ReturnRawString(const std::string& string)
-		{
-			char* rawString = (char*)malloc(string.size() + 1);
-			memcpy(rawString, string.c_str(), string.size());
-			rawString[string.size()] = '\0';
-			return rawString;
-		}
-
-		static void FreeSourceOnLoaded(WrenVM* vm, const char* name, struct WrenLoadModuleResult result)
-		{
-			(vm);
-			(name);
-			if (result.source != nullptr)
-				free((void*)result.source);
-		}
-
-		static const char* WrenResolveModule(
-			WrenVM* vm,
-			const char* importer,
-			const char* moduleName)
-		{
-			(vm);
-
-			// Logical import strings are used as-is and need no resolution.
-			if (moduleName == std::string_view(SoupModuleName))
-				return moduleName;
-			
-			// Get the directory containing the importing module.
-			auto modulePath = Path(importer);
-			auto moduleDirectory = modulePath.GetParent();
-
-			// Ensure the module is a relative path
-			auto moduleReference = Path(moduleName);
-			if (moduleReference.HasRoot())
-				return moduleName;
-
-			// Load the relative path from the current module folder
-			auto resolvedModule = moduleDirectory + moduleReference;
-			
-			// Automatically append wren file extension
-			resolvedModule.SetFileExtension("wren");
-
-			return ReturnRawString(resolvedModule.ToString());
-		}
-
-		static WrenLoadModuleResult WrenLoadModule(WrenVM* vm, const char* module)
-		{
-			(vm);
-			auto moduleName = std::string_view(module);
-			// Log::Diag("LoadModule: " + std::string(moduleName));
-
-			WrenLoadModuleResult result = {0};
-			result.onComplete = nullptr;
-
-			// Inject Soup module
-			if (moduleName == SoupModuleName)
-			{
-				result.source = GetSoupModuleSource();
-			}
-			else
-			{
-				// Attempt to load the module as a script file
-				std::ifstream scriptFile(module);
-				if (scriptFile.is_open())
-				{
-					auto script = std::string(
-						std::istreambuf_iterator<char>(scriptFile),
-						std::istreambuf_iterator<char>());
-						
-					result.onComplete = &FreeSourceOnLoaded;
-					result.source = ReturnRawString(script);
-				}
-			}
-
-			return result;
-		}
-
-		static void WrenWriteCallback(WrenVM* vm, const char* text)
-		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
-			host->WriteCallback(text);
-		}
-
-		static void WrenErrorCallback(
-			WrenVM* vm,
-			WrenErrorType errorType,
-			const char* module,
-			const int line,
-			const char* msg)
-		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
-			std::optional<std::string_view> moduleName = std::nullopt;
-			if (module != nullptr)
-				moduleName = module;
-			host->ErrorCallback(errorType, moduleName, line, msg);
-		}
-
 		static void SoupLoadGlobalState(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLoadGlobalState();
 		}
 
 		static void SoupLoadActiveState(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLoadActiveState();
 		}
 
 		static void SoupLoadSharedState(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLoadSharedState();
 		}
 
 		static void SoupCreateOperation(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupCreateOperation();
 		}
 
 		static void SoupLogInfo(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLogInfo();
 		}
 
 		static void SoupLogWarning(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLogWarning();
 		}
 
 		static void SoupLogError(WrenVM* vm)
 		{
-			auto host = (WrenHost*)wrenGetUserData(vm);
+			auto host = (GenerateHost*)wrenGetUserData(vm);
 			host->SoupLogError();
 		}
 
 		static const char* GetSoupModuleSource()
 		{
 			return soupModuleSource;
-		}
-
-		static WrenForeignMethodFn TryBindSoupModuleForeignMethod(
-			std::string_view className,
-			bool isStatic,
-			std::string_view signature)
-		{
-			// There is only one foreign method in the soup module.
-			if (className == SoupClassName && isStatic)
-			{
-				if (signature == "loadGlobalState_()")
-					return SoupLoadGlobalState;
-				else if (signature == "loadActiveState_()")
-					return SoupLoadActiveState;
-				else if (signature == "loadSharedState_()")
-					return SoupLoadSharedState;
-				else if (signature == "createOperation_(_,_,_,_,_,_)")
-					return SoupCreateOperation;
-				else if (signature == "info_(_)")
-					return SoupLogInfo;
-				else if (signature == "warning_(_)")
-					return SoupLogWarning;
-				else if (signature == "error_(_)")
-					return SoupLogError;
-			}
-
-			return nullptr;
 		}
 
 		static inline const char* soupModuleSource =
